pyproject.toml
[project]
name = "video2md"
version = "0.1.0"
description = "Convert videos into structured Markdown notes with ASR and keyframe understanding."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "video2md" }
]
dependencies = [
  "typer>=0.16.0",
  "pydantic>=2.11.2",
  "pyyaml>=6.0.2",
  "scenedetect[opencv]>=0.6.7.1",
  "rich>=14.0.0",
  "imageio-ffmpeg>=0.6.0",
  "httpx>=0.28.0"
]

[project.optional-dependencies]
asr = [
  "faster-whisper>=1.1.1,<1.2.0",
  "onnxruntime<1.19; python_full_version < '3.11'",
  "requests>=2.32.0"
]
ocr = [
  "paddleocr>=2.7.0,<3.0.0",
  "paddlepaddle>=2.6.0,<3.0.0",
  "rapidocr-onnxruntime>=1.4.4"
]
dev = [
  "pytest>=8.4.2",
  "ruff>=0.13.0"
]

[project.scripts]
video2md = "video2md.cli:app"

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/video2md"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
"src/video2md/cli.py" = ["B008"]
